requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
distlib==0.3.1
Django==3.1.7
django-crispy-forms==1.11.2
djangorestframework==3.12.4
filelock==3.0.12
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dotenv==0.16.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.3
wrapt==1.12.1
